requirements.txt
fastapi
requests
gradio_client
asyncio
websockets
uvicorn
pydub
python-multipart
aiohttp
anthropic
langchain
langchain_community
langchain_core
langchain_openai
langchain_mongodb
httpx_cache
passlib
langchainhub
bcrypt
PyJWT
boto3
chromadb
motor
emoji